requirements.txt
# For test_api.py - Google Gemini AI
google-generativeai==0.8.5
python-dotenv==1.2.1

# For listen.py - Speech recognition
SpeechRecognition==3.14.3
BanglaSpeech2Text==1.1.0
pyaudio==0.2.14
